requirements.txt
Flask
psycopg2-binary
flask-cors
gunicorn
